controlador horario, beneficio

+++ app/Http/Controllers/api/BeneficiosController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\api;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\beneficios;
+use Illuminate\Support\Facades\Auth;
 
 
 
@@ -25,7 +26,18 @@ class BeneficiosController extends Controller
      */
     public function store(Request $request)
     {
-        //
+        $user = Auth::user();
+
+        if (!$user || !$user->roles) {
+            return response()->json(['message' => 'Rol no definido'], 403);
+        }
+
+        if ($user->roles->Nombre_rol !== 'Administrador') {
+            return response()->json([
+                'message' => 'No tienes permisos para crear beneficios.'
+            ], 403);
+        }
+
         $beneficios = beneficios::create($request->all());
         return response()->json(['message' => 'Beneficio creado correctamente', 'beneficios' => $beneficios], 201);
     }
@@ -49,12 +61,21 @@ class BeneficiosController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        //
+        $user = Auth::user();
+
+        if (!$user || !$user->roles) {
+            return response()->json(['message' => 'Rol no definido'], 403);
+        }
+
+        if ($user->roles->Nombre_rol !== 'Administrador') {
+            return response()->json([
+                'message' => 'No tienes permisos para actualizar beneficios.'
+            ], 403);
+        }
+
         $beneficio = beneficios::find($id);
         if ($beneficio) {
-            //return response()->json(['beneficio' => $beneficio, 'request' => $request->all()]);
             $beneficio->update($request->all());
-
             return response()->json(['message' => 'Beneficio actualizado correctamente', 'beneficio' => $beneficio]);
         } else {
             return response()->json(['message' => 'Beneficio no encontrado'], 404);
@@ -66,7 +87,18 @@ class BeneficiosController extends Controller
      */
     public function destroy(string $id)
     {
-        //
+        $user = Auth::user();
+
+        if (!$user || !$user->roles) {
+            return response()->json(['message' => 'Rol no definido'], 403);
+        }
+
+        if ($user->roles->Nombre_rol !== 'Administrador') {
+            return response()->json([
+                'message' => 'No tienes permisos para eliminar beneficios.'
+            ], 403);
+        }
+
         $beneficio = beneficios::find($id);
         if ($beneficio) {
             $beneficio->delete();
